@@ -19,11 +19,11 @@
 from __future__ import unicode_literals
 
 import re
-from typing import Any, cast, Dict, List, Match, Optional
+from typing import Any, Dict, List, Match, Optional, cast
 
 from pkg_resources import Requirement as Req
 
-from .fragment import get_hash_info, parse_fragment, parse_extras_require
+from .fragment import get_hash_info, parse_extras_require, parse_fragment
 from .vcs import VCS, VCS_SCHEMES
 
 URI_REGEX = re.compile(
@@ -90,7 +90,7 @@ class Requirement:
         self.specs: List[str] = []
 
     def __repr__(self) -> str:
-        return '<Requirement: "{0}">'.format(self.line)
+        return f'<Requirement: "{self.line}">'
 
     def __getitem__(self, key: str) -> Any:
         return getattr(self, key)
@@ -128,7 +128,7 @@ class Requirement:
         :raises: ValueError on an invalid requirement
         """
 
-        req = cls('-e {0}'.format(line))
+        req = cls(f'-e {line}')
         req.editable = True
 
         if ' #' in line:
@@ -140,9 +140,9 @@ class Requirement:
         if vcs_match is not None:
             groups: Dict[str, str] = vcs_match.groupdict()
             if groups.get('login'):
-                req.uri = '{scheme}://{login}@{path}'.format(**groups)  # type: ignore
+                req.uri = f'{groups["scheme"]}://{groups["login"]}@{groups["path"]}'  # type: ignore
             else:
-                req.uri = '{scheme}://{path}'.format(**groups)  # type: ignore
+                req.uri = f'{groups["scheme"]}://{groups["path"]}'  # type: ignore
             req.revision = groups['revision']  # type: ignore
             if groups['fragment']:
                 fragment = parse_fragment(groups['fragment'])
@@ -188,9 +188,9 @@ class Requirement:
         if vcs_match is not None:
             groups = vcs_match.groupdict()
             if groups.get('login'):
-                req.uri = '{scheme}://{login}@{path}'.format(**groups)  # type: ignore
+                req.uri = f'{groups["scheme"]}://{groups["login"]}@{groups["path"]}'  # type: ignore
             else:
-                req.uri = '{scheme}://{path}'.format(**groups)  # type: ignore
+                req.uri = f'{groups["scheme"]}://{groups["path"]}'  # type: ignore
             req.revision = groups['revision']  # type: ignore
             if groups['fragment']:
                 fragment = parse_fragment(groups['fragment'])
@@ -203,7 +203,7 @@ class Requirement:
                     req.vcs = vcs  # type: ignore
         elif uri_match is not None:
             groups = uri_match.groupdict()
-            req.uri = '{scheme}://{path}'.format(**groups)  # type: ignore
+            req.uri = f'{groups["scheme"]}://{groups['path']}'  # type: ignore
             if groups['fragment']:
                 fragment = parse_fragment(groups['fragment'])
                 egg = fragment.get('egg')
